Move objects to their own file
